More tests

+++ spec/spec_helper.rb
@@ -6,16 +6,22 @@ require 'skylight'
 Dir[File.expand_path('../support/*.rb', __FILE__)].each { |f| require f }
 
 RSpec.configure do |config|
-  config.include SpecHelper::Path
+  config.include SpecHelper
 
   original_wd = Dir.pwd
 
   config.before :each do
-    FileUtils.rm_rf tmp
+    if File.exist?(tmp)
+      FileUtils.rm_rf tmp
+    end
   end
 
   config.after :each do
-    Dir.chdir(original_wd)
+    begin
+      cleanup_all_spawned_workers
+    ensure
+      Dir.chdir(original_wd)
+    end
   end
 
 end
